changed db from h2 to mysql

+++ src/main/java/ro/johann/thoughts/persistence/ThoughtRepository.java
@@ -7,6 +7,6 @@ import java.util.Optional;
 
 public interface ThoughtRepository {
     Thought create(Thought thought);
-    Optional<Thought> get(Integer id);
+    Optional<Thought> get(Long id);
     List<Thought> list();
 }
